Consolidate /users/me/identities into IdentityCommand, drop user_identities module
Moves ``list_from_user`` (the only retained method on the deprecated
UserIdentityCommand) onto IdentityCommand using
``self._client.url('users', 'me', 'identities')`` to reach the
non-/identities path — same pattern as wazo-auth-client's token
command. Removes the per-user CRUD methods (list / get / create /
update / delete) which are superseded by IdentityCommand on the new
flat /identities path.

The ``user_identities`` entry point is removed; consumers move to
``client.identities.list_from_user(...)``.
